Added static library VS solution to allow easy inclusion into larger projects. Updated FreeRTOS configuration macros.

+++ Chimera/sim/sim_chimera.cpp
@@ -14,17 +14,17 @@
 namespace ChimeraSim
 {
 
-  uint32_t millis()
+  size_t millis()
   {
     return 0u;
   }
 
-  void delayMilliseconds(uint32_t val )
+  void delayMilliseconds( size_t val )
   {
 
   }
 
-  void delayMicroseconds(uint32_t val )
+  void delayMicroseconds( size_t val )
   {
 
   }
